image-editor/requirements.txt
flask==2.3.3
pillow==10.4.0
requests==2.32.3